v3/lib: unify transforming output to args for compiledb generation
For example, instead of having one way to handle gcc commands with the
command line still in a string and another for the command line split,
always split.

This means that gccStrace and gccMakeX can be merged into a unified
handleGCC since they ultimately do the same thing.

There are some TODOs around making this better. A next step will be to
have the new handleGCC and similar methods return []cdbItem instead of
modifying w.it directly. This will allow them to return multiple
items, or none if the arguments are not interesting.